+++ tests/test_cuda_mem_get_info.py
@@ -132,3 +132,37 @@ def test_case_6():
         ["result"],
         reason="paddle does not support this function temporarily",
     )
+
+
+class CudaMemGetInfoAPIBase(APIBase):
+    def compare(
+        self,
+        name,
+        pytorch_result,
+        paddle_result,
+        check_value=True,
+        check_shape=True,
+        check_dtype=True,
+        check_stop_gradient=True,
+        rtol=1.0e-6,
+        atol=0.0,
+    ):
+        assert isinstance(pytorch_result, tuple) and len(pytorch_result) == 2
+        assert isinstance(paddle_result, tuple) and len(paddle_result) == 2
+
+
+cuda_mem_get_info_obj = CudaMemGetInfoAPIBase("torch.cuda.mem_get_info")
+
+
+@pytest.mark.skipif(
+    condition=not paddle.device.is_compiled_with_cuda(),
+    reason="can only run on paddle with CUDA",
+)
+def test_case_7():
+    pytorch_code = textwrap.dedent(
+        """
+        import torch
+        result = torch.cuda.mem_get_info()
+        """
+    )
+    cuda_mem_get_info_obj.run(pytorch_code, ["result"])
